Range check and size_t

+++ simplemd/LinkedCell.h
@@ -87,7 +87,7 @@ public:
 
   KOKKOS_INLINE_FUNCTION bool isGhostCell() const { return _isGhostCell; }
 
-  KOKKOS_INLINE_FUNCTION unsigned int getIndex() const { return _cellIndex; }
+  KOKKOS_INLINE_FUNCTION size_t getIndex() const { return _cellIndex; }
 
 private:
   KOKKOS_INLINE_FUNCTION void changeMoleculeCount(int by) { (*_linkedCellNumMolecules)(_cellIndex) += by; }
